requirements.txt
accelerate==0.25.0
agno==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncio==4.0.0
attrs==25.4.0
audioread==3.0.1
bertviz==1.4.1
boto3==1.39.4
botocore==1.39.4
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==2.0.12
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.2
cryptography==46.0.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.3.7
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.16
edge-tts==7.2.3
executing==2.2.0
fastapi==0.115.14
filelock==3.18.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
onnxruntime==1.22.0
packaging==25.0
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.8
playsound==1.3.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
pyparsing==3.2.3
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
setuptools==71.1.0
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.37.1
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
torchaudio==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
useful-moonshine-onnx @ git+https://git@github.com/usefulsensors/moonshine.git@89981f1ba33fc471ecb2e7a64531ee44f05e7de6#subdirectory=moonshine-onnx
uv==0.7.19
uvicorn==0.35.0
wcwidth==0.2.13
wheel==0.44.0
xxhash==3.5.0
yarl==1.22.0
tensorflow